pyproject.toml
[project]
name = "resume"
version = "0.1.0"
description = "HTML to PDF conversion"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docopt-ng>=0.9.0",
    "pytest-playwright==0.7.0",
    "pytest-playwright-asyncio==0.7.0",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
